requirements.txt
setuptools
requests
pandas
pyyaml
deepl
